+++ packages/extension-store-locator/src/hooks/use-extension-config.ts
@@ -13,5 +13,8 @@ import extensionMeta from '../../extension-meta.json'
  */
 export const useExtensionConfig = () => {
     const extension = useApplicationExtension(extensionMeta.id)
+    if (extension === undefined) {
+        throw new Error(`'useExtensionConfig' could find your current application extension instance!`)
+    }
     return extension?.getConfig()
 }
